complete work02

+++ src/template/work00.rs
@@ -4,14 +4,14 @@ use std::{fs, cmp};
 use rand::Rng;
 
 /// Converts an image into the number spiral gradient
-pub fn number_spiral(img : &mut Image, scale1: f64, scale2: f64) {
+pub fn number_spiral(img : &mut Image, scale1: f32, scale2: f32) {
     // Looping through pixels
     for i in 0..img.get_height() {
         for q in 0..img.get_width() {
             // Initializing rgb colors
             let r : u8 = 0;
-            let g : u8 = (number_spiral_helper(((i as f64) * scale1) as i32, ((q as f64) * scale1) as i32) % 256) as u8;
-            let b : u8 = (number_spiral_helper(((q as f64) * scale2) as i32, ((i as f64) * scale2) as i32) % 256) as u8;
+            let g : u8 = (number_spiral_helper(((i as f32) * scale1) as i32, ((q as f32) * scale1) as i32) % 256) as u8;
+            let b : u8 = (number_spiral_helper(((q as f32) * scale2) as i32, ((i as f32) * scale2) as i32) % 256) as u8;
 
             // Updating pixel
             img.update_pixel_rc(i, q, Pixel::new_value(r, g, b));
@@ -44,14 +44,14 @@ fn number_spiral_helper(row : i32, col : i32) -> i32 {
 }
 
 /// Converts an image into the number grid gradient
-pub fn number_grid(img : &mut Image, scale1 : f64, scale2 : f64) {
+pub fn number_grid(img : &mut Image, scale1 : f32, scale2 : f32) {
     // Looping through pixels
     for i in 0..img.get_height() {
         for q in 0..img.get_width() {
             // Initializing rgb colors
             let r : u8 = 0;
-            let g : u8 = (number_grid_helper(((i as f64) * scale1) as i32, ((q as f64) * scale1) as i32) % 256) as u8;
-            let b : u8 = (number_grid_helper(((q as f64) * scale2) as i32, ((i as f64) * scale2) as i32) % 256) as u8;
+            let g : u8 = (number_grid_helper(((i as f32) * scale1) as i32, ((q as f32) * scale1) as i32) % 256) as u8;
+            let b : u8 = (number_grid_helper(((q as f32) * scale2) as i32, ((i as f32) * scale2) as i32) % 256) as u8;
 
             // Updating pixel
             img.update_pixel_rc(i, q, Pixel::new_value(r, g, b));
@@ -66,14 +66,14 @@ fn number_grid_helper(row : i32, col : i32) -> i32 {
 }
 
 /// Converts an image into a counting bits gradient
-pub fn counting_bits(img : &mut Image, scale1 : f64, scale2 : f64) {
+pub fn counting_bits(img : &mut Image, scale1 : f32, scale2 : f32) {
     // Looping through pixels
     for i in 0..img.get_height() {
         for q in 0..img.get_width() {
             // Initializing rgb colors
             let r : u8 = 0;
-            let g : u8 = (counting_bits_helper((((i^q) as f64) * scale1) as i64) % 256) as u8;
-            let b : u8 = (counting_bits_helper((((i^q) as f64) * scale2) as i64) % 256) as u8;
+            let g : u8 = (counting_bits_helper((((i^q) as f32) * scale1) as i64) % 256) as u8;
+            let b : u8 = (counting_bits_helper((((i^q) as f32) * scale2) as i64) % 256) as u8;
 
             // Updating pixel
             img.update_pixel_rc(i, q, Pixel::new_value(r, g, b));
@@ -87,7 +87,7 @@ fn counting_bits_helper(n : i64) -> i64 {
     let mut ans : i64 = 0;
 
     // Looping through bits
-    for i in 0..((n as f64).log2() as i32) {
+    for i in 0..((n as f32).log2() as i32) {
         // Variable declarations
         let div : i64 = counting_bits_modpow(2, i as i64);
         let mut cnt : i64 = (n+1)/div;
@@ -122,11 +122,11 @@ fn counting_bits_modpow(x : i64, n : i64) -> i64 {
 }
 
 /// Function that creates the barnsley fern on an image
-pub fn barnsley(img : &mut Image, pix : Pixel, itr : i32, scale : f64) {
+pub fn barnsley(img : &mut Image, pix : Pixel, itr : i32, scale : f32) {
     // Setting up random generator and array
     let mut rng = rand::thread_rng();
-    let mut xval : Vec<f64> = vec![0.0; itr as usize];
-    let mut yval : Vec<f64> = vec![0.0; itr as usize];
+    let mut xval : Vec<f32> = vec![0.0; itr as usize];
+    let mut yval : Vec<f32> = vec![0.0; itr as usize];
 
     // Looping through iterations
     for i in 1..itr {
@@ -141,15 +141,15 @@ pub fn barnsley(img : &mut Image, pix : Pixel, itr : i32, scale : f64) {
         } else if (prob <= 86) {
             // The small leaflet
             xval[i as usize] = ((0.85 * (xval[(i-1) as usize])) + (0.04 * (yval[(i-1) as usize])));
-            yval[i as usize] = ((-0.04 * (xval[(i-1) as usize])) + (0.85 * (yval[(i-1) as usize])) + (img.get_width() as f64)/scale);
+            yval[i as usize] = ((-0.04 * (xval[(i-1) as usize])) + (0.85 * (yval[(i-1) as usize])) + (img.get_width() as f32)/scale);
         } else if (prob <= 93) {
             // The large left leaflet
             xval[i as usize] = ((0.2 * (xval[(i-1) as usize])) - (0.26 * (yval[(i-1) as usize])));
-            yval[i as usize] = ((0.23 * (xval[(i-1) as usize])) + (0.22 * (yval[(i-1) as usize])) + (img.get_width() as f64)/scale);
+            yval[i as usize] = ((0.23 * (xval[(i-1) as usize])) + (0.22 * (yval[(i-1) as usize])) + (img.get_width() as f32)/scale);
         } else {
             // The large right leaflet
             xval[i as usize] = ((-0.15 * (xval[(i-1) as usize])) + (0.28 * (yval[(i-1) as usize])));
-            yval[i as usize] = ((0.26 * (xval[(i-1) as usize])) + (0.24 * (yval[(i-1) as usize])) + (img.get_width() as f64)/(scale*4.0));
+            yval[i as usize] = ((0.26 * (xval[(i-1) as usize])) + (0.24 * (yval[(i-1) as usize])) + (img.get_width() as f32)/(scale*4.0));
         }
     }
 
@@ -157,7 +157,7 @@ pub fn barnsley(img : &mut Image, pix : Pixel, itr : i32, scale : f64) {
     for i in 0..itr {
         // Plotting points
         let width : i32 = img.get_width();
-        img.update_pixel_xy((xval[i as usize] + ((width as f64/(scale/3.3)))) as i32, yval[i as usize] as i32, pix)
+        img.update_pixel_xy((xval[i as usize] + ((width as f32/(scale/3.3)))) as i32, yval[i as usize] as i32, pix)
     }
 }
 
@@ -171,17 +171,17 @@ pub fn create_work00_images() {
 
     // Number spiral image
     let mut curr1 : Image = Image::new_dimension(size, size);
-    number_spiral(&mut curr1, 0.2 as f64, 0.2 as f64);
+    number_spiral(&mut curr1, 0.2 as f32, 0.2 as f32);
     file::create_ppm_ascii("image/ppm/w00_corridor.ppm", curr1);
 
     // Number grid image
     let mut curr2 : Image = Image::new_dimension(size, size);
-    number_grid(&mut curr2, 1.0 as f64, 0.5 as f64);
+    number_grid(&mut curr2, 1.0 as f32, 0.5 as f32);
     file::create_ppm_ascii("image/ppm/w00_checkerboard.ppm", curr2);
 
     // Counting bits image
     let mut curr3 : Image = Image::new_dimension(size, size);
-    counting_bits(&mut curr3, 0.6 as f64, 0.6 as f64);
+    counting_bits(&mut curr3, 0.6 as f32, 0.6 as f32);
     file::create_ppm_ascii("image/ppm/w00_chains.ppm", curr3);
 
     // Barnsley fern drawing
